+++ src/sbm_gwf_model.jl
@@ -348,13 +348,12 @@ function initialize_sbm_gwf_model(config::Config)
     )
     writer = prepare_writer(
         config,
-        reader,
         modelmap,
         indices_reverse,
         x_nc,
         y_nc,
         nc,
-        maxlayers = sbm.maxlayers,
+        extra_dim = (name = "layer", value = Float64.(1:sbm.maxlayers)),
     )
     close(nc)